Stop partial menu generation in INSTALL file
The commit d136c6dc resulted in menu text for the "Top" node being added
to the INSTALL file on regeneration.  As the full menu is not displayed
in the plain text file anyway, suppress the menu section completely to
avoid the additional text.

Also regenerate the INSTALL file to commit a small formatting change
introduced in the same commit.