Move shared umount.c from hppa to mips.
This patch prepares for moving mips from ports to libc by reversing
the #include ordering between mips and hppa.

Reversing #include ordering for umount.c is the conservative change in
preparation for moving the mips port.  In fact, it appears there are
several redundant umount.c files all implementing umount in terms of
the umount2 syscall; I've filed bug 16552 for eliminating that
redundancy properly.

Tested that disassembly of shared libraries for mips (o32, n32, n64)
is the same before and after this patch.

	* sysdeps/unix/sysv/linux/hppa/umount.c: Move to
	sysdeps/unix/sysv/linux/mips/mips64/umount.c and #include that
	file.

	* sysdeps/unix/sysv/linux/mips/mips64/umount.c: Move from
	sysdeps/unix/sysv/linux/hppa/umount.c instead of #include of that
	file.