Remove reference to subdir_lint.out.
The toplevel Makefile's subdir_targets variable refers to
subdir_lint.out.  As far as I can tell, this is not defined or
referenced anywhere else in the tree.  Having .out makefile references
that don't refer to the output from testcases seems confusing; this
patch removes the reference to subdir_lint.out.

Tested x86_64.

	* Makefile (subdir_targets): Remove subdir_lint.out.